+++ dec26/.gitignore
